+++ web/package.json
@@ -1,29 +1,26 @@
 {
-  "type": "module",
-  "author": "Tiffany Bennett (https://tiffnix.com)",
-  "name": "rink-web",
-  "version": "0.6.3",
-  "repository": "https://github.com/tiffany352/rink-rs",
-  "homepage": "https://rinkcalc.app",
-  "description": "Web interface to Rink",
-  "license": "MPL-2.0",
-  "devDependencies": {
-    "@types/elementtree": "^0.1.0",
-    "@types/node-fetch": "^2.5.7",
-    "@sveltejs/kit": "^2.5.5",
-    "svelte": "^4.2.14",
-    "svelte-preprocess": "^5.1.3",
-    "tslib": "^2.0.0",
-    "typescript": "^5.4.5",
-    "asciidoctor": "^3"
-  },
-  "dependencies": {
-    "elementtree": "^0.1.7",
-    "node-fetch": "^2.6.1",
-    "atomically": "1.7.0"
-  },
-  "browserslist": [
-    "last 3 chrome versions",
-    "last 3 firefox versions"
-  ]
+	"type": "module",
+	"author": "Tiffany Bennett (https://tiffnix.com)",
+	"name": "rink-web",
+	"version": "0.6.3",
+	"repository": "https://github.com/tiffany352/rink-rs",
+	"homepage": "https://rinkcalc.app",
+	"description": "Web interface to Rink",
+	"license": "MPL-2.0",
+	"devDependencies": {
+		"@types/elementtree": "^0.1.0",
+		"@types/node-fetch": "^2.5.7",
+		"tslib": "^2.0.0",
+		"typescript": "^5.4.5",
+		"asciidoctor": "^3"
+	},
+	"dependencies": {
+		"elementtree": "^0.1.7",
+		"node-fetch": "^2.6.1",
+		"atomically": "1.7.0"
+	},
+	"browserslist": [
+		"last 3 chrome versions",
+		"last 3 firefox versions"
+	]
 }
