--- web/src/components/expr/Call.svelte
@@ -1,12 +1,0 @@
-<script lang="typescript">
-  import type { CallExpr } from "../../util/expr";
-
-  export let value: CallExpr;
-</script>
-
-<span>{value.func}</span>
-<span>(</span>
-{#each value.args as arg}
-  <svelte:self value={arg} />
-{/each}
-<span>)</span>
