--- web/src/components/expr/Const.svelte
@@ -1,10 +1,0 @@
-<script lang="typescript">
-  import type { ConstExpr } from "../../util/expr";
-  import Numeric from "../Numeric.svelte";
-
-  export let value: ConstExpr;
-</script>
-
-<span>
-  <Numeric number={value.value} />
-</span>
