--- web/src/components/expr/Date.svelte
@@ -1,7 +1,0 @@
-<script lang="typescript">
-  import type { DateExpr } from "../../util/expr";
-
-  export let value: DateExpr;
-</script>
-
-<pre>{(JSON.stringify(value), null, 2)}</pre>
