--- web/src/components/expr/Expr.svelte
@@ -1,10 +1,0 @@
-<script lang="typescript">
-  import Node from "./Node.svelte";
-  import type { Expr } from "../../util/expr";
-
-  export let value: Expr;
-</script>
-
-<span>
-  <Node {value} />
-</span>
