--- web/src/components/expr/Quote.svelte
@@ -1,7 +1,0 @@
-<script lang="typescript">
-  import type { QuoteExpr } from "../../util/expr";
-
-  export let value: QuoteExpr;
-</script>
-
-<span>"{value.string}"</span>
