--- web/src/components/expr/Unaryop.svelte
@@ -1,9 +1,0 @@
-<script lang="typescript">
-  import type { UnaryOpExpr } from "../../util/expr";
-  import Node from "./Node.svelte";
-
-  export let value: UnaryOpExpr;
-</script>
-
-<span>{value.op}</span>
-<Node value={value.expr} />
