--- web/src/components/expr/Unit.svelte
@@ -1,7 +1,0 @@
-<script lang="typescript">
-  import type { UnitExpr } from "../../util/expr";
-
-  export let value: UnitExpr;
-</script>
-
-<a href={`/unit/${value.name}`}>{value.name}</a>
