--- web/src/components/reply/ConversionReply.svelte
@@ -1,10 +1,0 @@
-<script lang="typescript">
-  import type { ConversionReply } from "../../util/reply";
-  import Number from "../Number.svelte";
-
-  export let value: ConversionReply;
-</script>
-
-<p>
-  <Number number={value.value} />
-</p>
