--- web/src/components/reply/FactorizeReply.svelte
@@ -1,32 +1,0 @@
-<script lang="typescript">
-  import type { FactorizeReply } from "../../util/reply";
-
-  export let value: FactorizeReply;
-</script>
-
-<style>
-  li > div {
-    display: inline-flex;
-    flex-direction: row;
-  }
-
-  .space {
-    width: 0.25em;
-  }
-</style>
-
-<ul>
-  {#each value.factorizations as unit}
-    <li>
-      <div>
-        {#each Object.entries(unit) as [dim, pow], i}
-          {#if i != 0}
-            <span class="space">&nbsp;</span>
-          {/if}
-          <span>{dim}</span>
-          <sup>{pow}</sup>
-        {/each}
-      </div>
-    </li>
-  {/each}
-</ul>
