--- web/src/components/reply/GenericError.svelte
@@ -1,7 +1,0 @@
-<script lang="typescript">
-  import type { GenericError } from "../../util/reply";
-
-  export let value: GenericError;
-</script>
-
-<p>{value.message}</p>
