--- web/src/components/reply/NumberReply.svelte
@@ -1,10 +1,0 @@
-<script lang="typescript">
-  import type { NumberReply } from "../../util/reply";
-  import Number from "../Number.svelte";
-
-  export let value: NumberReply;
-</script>
-
-<p>
-  <Number number={value} />
-</p>
