--- web/src/routes/query/index.svelte
@@ -1,7 +1,0 @@
-<script context="module">
-  export async function preload(page, session) {
-    let { input } = page.query;
-    const slug = encodeURIComponent(input || "");
-    return this.redirect(302, `/query/${slug}`);
-  }
-</script>
