--- web/svelte.config.js
@@ -1,5 +1,0 @@
-import sveltePreprocess from "svelte-preprocess";
-
-module.exports = {
-	preprocess: sveltePreprocess({}),
-};
